Functions

Behavior.runspec — Function

runspec(rootpath; featurepath, stepspath, execenvpath, parseoptions, presenter, tags)

Execute all features found from the rootpath.

By default, it looks for feature files in <rootpath>/features and step files <rootpath>/features/steps. An environment.jl file may be added to <rootpath>/features directory for running certain before/after code. You may override the default locations by specifying featurepath, stepspath, or execenvpath.

The tagselector option is an expression you can use to select which scenarios to run based on tags. For instance, the tag selector @foo will run only those scenarios that have the tag @foo, while not @ignore will run only that scenarios that do not have the @ignore tag.

See also: Gherkin.ParseOptions.

Behavior.Gherkin.ParseOptions — Type

ParseOptions lets the user control certain behavior of the parser, making it more lenient towards errors.

The "experimental" Gherkin parser in GherkinExperimental.jl is now the default parser, as it surpassed the older parser in many ways.

Index

Behavior.Gherkin.ParseOptions

Behavior.runspec